+++ src/shared/components/idme/idme.css
@@ -1,5 +1,6 @@
 .wrapper {
 	text-align: center;
+	margin: 2%;
 }
 
 .auth {
